Adding all project files to the repository
read_pressure_sensor.c is the main file for the project and its entry point. run_dockerized_solution.sh is a bash script that automates the deployment of the desktop sandbox for verifying the functionality of the pressure sensor driver, using the Dockerfile as the base instructions for building the image.